+++ photons_messages/fields.py
@@ -100,7 +100,8 @@ tile_state_device = (
     , ("device_version_version", T.Uint32)
     , ("firmware_build", T.Uint64)
     , ("reserved8", T.Reserved(64))
-    , ("firmware_version", T.Uint32.version_number())
+    , ("firmware_version_minor", T.Uint16)
+    , ("firmware_version_major", T.Uint16)
     , ("reserved9", T.Reserved(32))
     )